- Fixed layout of program - Renamed timesheet entry & details page
