+++ app/src/main/java/com/yugen/opsc7311_poe/LoginPageActivity.kt
@@ -3,6 +3,8 @@ package com.yugen.opsc7311_poe
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.yugen.opsc7311_poe.databinding.LoginPageBinding
+import com.yugen.opsc7311_poe.helpers.openIntent
+
 class LoginPageActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
