+++ app/src/main/java/com/yugen/opsc7311_poe/helpers/IntentHelper.kt
@@ -1,4 +1,4 @@
-package com.yugen.opsc7311_poe
+package com.yugen.opsc7311_poe.helpers
 
 import android.app.Activity
 import android.content.Intent
